Implement INT 8 (#26)

+++ MachineEmulator/Operations/ArithmeticOperations.cs
@@ -5,17 +5,17 @@ namespace MachineEmulator.Operations
     class ArithmeticOperations
     {
         public static void ADD(Processor proc, RAM ram, Register reg1, Register reg2) {
-            long result = proc.registers[(int)reg1] + proc.registers[(int)reg2];
+            long result = (int)proc.registers[(int)reg1] + (int)proc.registers[(int)reg2];
             FlagUtils.AdjustResultFlags(proc, result);
             proc.registers[(int)reg1] = (uint)result;
         }
         public static void SUB(Processor proc, RAM ram, Register reg1, Register reg2) {
-            long result = proc.registers[(int)reg1] - proc.registers[(int)reg2];
+            long result = (int)proc.registers[(int)reg1] - (int)proc.registers[(int)reg2];
             FlagUtils.AdjustResultFlags(proc, result);
             proc.registers[(int)reg1] = (uint)result;
         }
         public static void MUL(Processor proc, RAM ram, Register reg1, Register reg2) {
-            long result = proc.registers[(int)reg1] * proc.registers[(int)reg2];
+            long result = (int)proc.registers[(int)reg1] * (int)proc.registers[(int)reg2];
             FlagUtils.AdjustResultFlags(proc, result);
             proc.registers[(int)Register.R1] = (uint)(result / uint.MaxValue);
             proc.registers[(int)Register.R2] = (uint)(result % uint.MaxValue);
@@ -25,12 +25,12 @@ namespace MachineEmulator.Operations
                 throw new NotImplementedException(); //TODO: IMPLEMENT DIV IS NULIO
                 return;
             }
-            long result = proc.registers[(int)reg1] / proc.registers[(int)reg2];
+            long result = (int)proc.registers[(int)reg1] / (int)proc.registers[(int)reg2];
             FlagUtils.AdjustResultFlags(proc, result);
             proc.registers[(int)reg1] = (uint)result;
         }
         public static void CMP(Processor proc, RAM ram, Register reg1, Register reg2) {
-            long result = proc.registers[(int)reg1] - proc.registers[(int)reg2];
+            long result = (int)proc.registers[(int)reg1] - (int)proc.registers[(int)reg2];
             FlagUtils.AdjustResultFlags(proc, result);
         }
     }
